Extracted filterDataItems
