+++ src/Data/localStorageDataConnection.ts
@@ -3,6 +3,7 @@ import { pseudoRandom32CharHex } from '../utils/pseudoRandom32CharHex'
 import { orderBy } from 'lodash-es'
 import { ensureString } from '../utils/ensureString'
 import { scrivitoTenantId } from '../config/scrivitoTenants'
+import { filterDataItems } from './filterDataItems'
 
 interface RawDataItem {
   _id: string
@@ -37,21 +38,7 @@ export function localStorageDataConnection(
         ? await Promise.all(rawItems.map((item) => postProcessData(item)))
         : rawItems
 
-      const filters = params.filters()
-      const filteredItems =
-        Object.keys(filters).length === 0
-          ? items
-          : items.filter((item) =>
-              Object.entries(filters).every(([filterAttribute, filter]) => {
-                const itemValue = item[filterAttribute]
-                const subFilters =
-                  filter.operator === 'and' ? filter.value : [filter]
-
-                return subFilters.every(({ value: filterValue, opCode }) =>
-                  compare({ itemValue, filterValue, opCode }),
-                )
-              }),
-            )
+      const filteredItems = filterDataItems(params.filters(), items)
 
       const search = params.search().toLowerCase()
       const matchingItems = filteredItems.filter((item) =>
@@ -212,48 +199,4 @@ function orderItems(
     order.map(([attr]) => attr),
     order.map(([_, ascOrDesc]) => ascOrDesc),
   )
-}
-
-const comparators = {
-  gt: (a: string | number, b: string | number) => a > b,
-  lt: (a: string | number, b: string | number) => a < b,
-  gte: (a: string | number, b: string | number) => a >= b,
-  lte: (a: string | number, b: string | number) => a <= b,
-}
-
-function compare({
-  itemValue,
-  filterValue,
-  opCode,
-}: {
-  itemValue: unknown
-  filterValue: string | number | boolean | null
-  opCode: 'eq' | 'neq' | 'gt' | 'lt' | 'gte' | 'lte'
-}): boolean {
-  if (opCode !== 'eq' && opCode !== 'neq') {
-    if (filterValue === null) return false
-
-    if (
-      !(
-        (typeof itemValue === 'number' || typeof itemValue === 'string') &&
-        (typeof filterValue === 'number' || typeof filterValue === 'string')
-      )
-    ) {
-      throw new Error(
-        `Invalid comparison: ${JSON.stringify(itemValue)} and ${JSON.stringify(filterValue)} must be numbers or strings.`,
-      )
-    }
-
-    return comparators[opCode](itemValue, filterValue)
-  }
-
-  const eq =
-    itemValue === filterValue ||
-    (filterValue === null && itemValue === undefined) ||
-    (filterValue === 'null' && itemValue === null) ||
-    (filterValue === 'null' && itemValue === undefined) ||
-    (filterValue === 'true' && itemValue === true) ||
-    (filterValue === 'false' && itemValue === false)
-
-  return opCode === 'neq' ? !eq : eq
 }
